+++ config/getcandy.php
@@ -48,6 +48,11 @@ return [
     |
     */
     'orders' => [
+        /**
+         * The invoice reference prefix for an order
+         * e.g {prefix}2019-04-15
+         */
+        'reference_prefix' => null,
         'mailers' => [
             // 'dispatched' => \Your\OrderDispatchedMailer::class,
             // 'payment-processing' => \Your\ConfirmationMailer::class,